+++ data_safe_haven/utility/logger.py
@@ -4,7 +4,6 @@ import logging
 from typing import Any
 
 from rich.logging import RichHandler
-from rich.table import Table
 from rich.text import Text
 
 from .singleton import Singleton
@@ -102,28 +101,6 @@ class LoggingSingleton(logging.Logger, metaclass=Singleton):
             return self.debug(remainder)
         else:
             return self.info(message.strip())
-
-    def tabulate(
-        self, header: list[str] | None = None, rows: list[list[str]] | None = None
-    ) -> list[str]:
-        """Generate a table from header and rows
-
-        Args:
-            header: The table header
-            rows: The table rows
-
-        Returns:
-            A list of strings representing the table
-        """
-        table = Table()
-        if header:
-            for item in header:
-                table.add_column(item)
-        if rows:
-            for row in rows:
-                table.add_row(*row)
-        adaptor = RichStringAdaptor(coloured=True)
-        return [line.strip() for line in adaptor.to_string(table).split("\n")]
 
 
 class NonLoggingSingleton(logging.Logger, metaclass=Singleton):
